requirements.txt
pip==21.3.1
python==3.6.15
python_abi==3.6
setuptools==58.0.4
ucrt==10.0.20348.0
vc==14.2
vs2015_runtime==14.29.30037
wheel==0.37.1
pytest==4.5.0
allure-pytest==2.8.6
PyYAML==5.4.1
requests==2.27.1